requirements.txt
django
GDAL==3.0.4
psycopg2
djangorestframework
djangorestframework-gis
